enable roles on order service and adding all needed endpoints

+++ product-service/src/main/java/com/amrfawzi/ecommerce/product_service/dto/UpdateProductRequest.java
@@ -13,7 +13,7 @@ public class UpdateProductRequest {
     private String description;
 
     @Min(0)
-    private BigDecimal price;
+    private Double price;
 
     @Min(0)
     private Integer quantity;
